fix: move unit test of Matrix creation into DecideTest.java
close #70
